backend2/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymongo==4.7.2
python-dotenv==1.0.1
paramiko==3.4.0
pydantic==2.8.2
psutil==6.0.0
